refactor to reduce complexity around apiErrorMap middleware
