+++ apps/google-analytics-4/lambda/src/middlewares/apiErrorHandler.spec.ts
@@ -3,7 +3,7 @@ import Express from 'express';
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
 import sinon from 'sinon';
-import { apiErrorHandler, ApiErrorMap, apiErrorMapper } from './apiErrorHandler';
+import { ApiErrorMap, apiErrorHandler, apiErrorMapper } from './apiErrorHandler';
 import { ApiError } from '../errors/apiError';
 
 describe('apiErrorHandler', () => {
